# ansi: speed up parsing by roughly 7.5x