This commit speeds up the parsing/processing of ANSI escape codes by roughly 7.5x. The speedup is mostly accomplished by replacing the regex with dedicated parsing logic ([`nextAnsiEscapeSequence()`](https://github.com/charlievieth/fzf/blob/3d4a123fa2c7f4daf9dc822fbc3e52214172d106/src/ansi.go#L122-L184)) and reducing the number of allocations in [`extractColor()`](https://github.com/charlievieth/fzf/blob/3d4a123fa2c7f4daf9dc822fbc3e52214172d106/src/ansi.go#L186-L276).

#### Benchmarks
```
name             old time/op    new time/op     delta
ExtractColor-16    4.89µs ± 5%     0.64µs ± 2%   -86.87%  (p=0.000 n=9+9)

name             old speed      new speed       delta
ExtractColor-16  25.6MB/s ± 5%  194.6MB/s ± 2%  +661.43%  (p=0.000 n=9+9)

name             old alloc/op   new alloc/op    delta
ExtractColor-16    1.37kB ± 0%     0.31kB ± 0%   -77.31%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op   delta
ExtractColor-16      48.0 ± 0%        4.0 ± 0%   -91.67%  (p=0.000 n=10+10)
```

#### TODO
* ANSI performance can be further improved by reducing conversions between `string` and `[]byte`, for example most callers of `extractColor` convert the string response to `[]byte`:
   * [`core.go:72:74`](https://github.com/junegunn/fzf/blob/29851c18aa47ad4b4cbf92358cfbe11935f0c033/src/core.go#L72-L74)
   * [`core.go:79:81`](https://github.com/junegunn/fzf/blob/29851c18aa47ad4b4cbf92358cfbe11935f0c033/src/core.go#L79-L81)
   * [`terminal.go:539:540`](https://github.com/junegunn/fzf/blob/29851c18aa47ad4b4cbf92358cfbe11935f0c033/src/terminal.go#L539-L540)
   * [`terminal.go:1021:1024`](https://github.com/junegunn/fzf/blob/29851c18aa47ad4b4cbf92358cfbe11935f0c033/src/terminal.go#L1021-L1024)