@junegunn 2021-02-25
Thanks, I'll review the changes when I get some time. Regarding TODO items, let's not try to optimize non-critical paths. We need to find a balance between performance and code readability/maintainability.