@junegunn 2021-02-28
I can confirm this makes `--ansi` significantly faster 👍 

[ponies.txt](https://github.com/junegunn/fzf/files/6056775/ponies.txt)

```sh
$ cat ponies.txt{,,,,,,,,,,,,,,,,,,,,,,} | fzf --ansi --reverse

$ cat ponies.txt{,,,,,,,,,,,,,,,,,,,,,,} | time fzf-without-patch --ansi -f x
        3.50 real         4.60 user         0.18 sys
$ cat ponies.txt{,,,,,,,,,,,,,,,,,,,,,,} | time fzf-with-patch --ansi -f x
        1.03 real         2.08 user         0.11 sys
```